requirements.txt
crewai==0.28.8
crewai-tools==0.1.6
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
python-multipart==0.0.6
jinja2==3.1.2
python-dotenv==1.0.0
openai==1.3.7
langchain==0.1.0
langchain-openai==0.0.2
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2
pandas==2.1.4
numpy==1.24.3
python-docx==1.1.0
PyPDF2==3.0.1

